+++ api_test/testutil_test.go
@@ -210,7 +210,7 @@ func NewTestServer(t testing.TB, ctx context.Context, blocking bool) *TestServer
 
 	sqlStore := sqlstore.SQLStore{}
 	sqlBalanceStore := sqlstore.NewBalances(&sqlStore)
-
+	sqlOrderStore := sqlstore.NewOrders(&sqlStore)
 	eventSource, err := broker.NewEventSource(conf.Broker, logger)
 
 	if err != nil {
@@ -270,6 +270,7 @@ func NewTestServer(t testing.TB, ctx context.Context, blocking bool) *TestServer
 		stakingService,
 		checkpointSvc,
 		sqlBalanceStore,
+		sqlOrderStore,
 	)
 	if srv == nil {
 		t.Fatal("failed to create gRPC server")