+++ cmd/data-node/node/node.go
@@ -105,6 +105,7 @@ type NodeCommand struct {
 	balanceStoreSQL *sqlstore.Balances
 	ledgerSQL       *sqlstore.Ledger
 	partyStoreSQL   *sqlstore.Parties
+	orderStoreSQL   *sqlstore.Orders
 	tradeStoreSQL   *sqlstore.Trades
 
 	vegaCoreServiceClient vegaprotoapi.CoreServiceClient
@@ -138,6 +139,7 @@ type NodeCommand struct {
 	assetSubSQL            *sqlsubscribers.Asset
 	timeSubSQL             *sqlsubscribers.Time
 	transferResponseSubSQL *sqlsubscribers.TransferResponse
+	orderSubSQL            *sqlsubscribers.Order
 	tradesSubSQL           *sqlsubscribers.TradeSubscriber
 
 	candleService     *candles.Svc
@@ -238,6 +240,7 @@ func (l *NodeCommand) runNode(args []string) error {
 		l.stakingService,
 		l.checkpointSvc,
 		l.balanceStoreSQL,
+		l.orderStoreSQL,
 	)
 
 	// watch configs