+++ entities/party.go
@@ -2,6 +2,7 @@ package entities
 
 import (
 	"encoding/hex"
+	"fmt"
 	"time"
 )
 
@@ -10,6 +11,14 @@ type Party struct {
 	VegaTime time.Time
 }
 
-func (p Party) HexId() string {
+func (p Party) HexID() string {
 	return hex.EncodeToString(p.ID)
+}
+
+func MakePartyID(stringID string) ([]byte, error) {
+	id, err := hex.DecodeString(stringID)
+	if err != nil {
+		return nil, fmt.Errorf("party id is not valid hex string: %v", stringID)
+	}
+	return id, nil
 }